backend/foodgram/requirements.txt
atomicwrites==1.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==3.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.1.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
psycopg2==2.9.10
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
requests-oauthlib==2.0.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
toml==0.10.2
uritemplate==4.1.1
